定义了一些基础方法
[Feature] 兼容下标访问，并加入 Optional 防止 NilException
